search_continents  process_event
I created a function that takes a connection and event as a parameter and I create a sql statement which I then return the fetchall of the results. Which is a list of the data. Where In the process_events if the event matches it, I unpack values in the returned list and yield that into the new event. I hae to create the search_continents still though.