def new_country(connection, event) in country_events.py
I followed the same steps for my new_continents function but this time I created more if statements for NOT NULL columns since this table has more columns. Which basically if it is NOT NULL it cahnges it into empty strings But this case has keywords as NULL so i created 2 more if statements for if the keywords is none or not none which is either null or not in the sql statement.