def search_region(connection, events)
I did the same thing I did basically with the country_events and continent_events with the same parameters. But this time of course I had to change the parameters and there is more search boxes this times. So i needed more if statements to check different combinations of local code, region code, and name.