def edit_region(connection, events) + changing sql statements in continent_events
I created a new function following what I did for country_Events and continent_events but this time using I edited the sql statements so it is matching the correct table. I also edited the continent_events functions so that it provents injection attacks. By making the sql statements parameterized